Complete fourth section and modify other elements
The last section before the footer was done, completing the full page. In other elements, changes were made in the way that they were positioned, replacing "justify-content" and "align-items" with a simple "margin: auto" when it was possible. Unnecessary lines were removed also, like a margin/padding 0 in the quote section by declaring it universally (*).