Polish css and modify first section
Height for the first section was defined to 75vh, as it wasn't declared before. This was a final revision, enhancing code clarity and removing other unnecessary lines, such as margin/padding 0 that were found, since in the last commit a universal declaration of margin/padding was included. There was a flex: auto in the main section included in the first versions to organize header, main and footer that by this version wasn't necessary anymore.